fix: 답변 작성 시 답변 개수를 DB에 저장하는 방식을 redis를 사용하는 방식으로 변경

+++ src/main/java/com/hong/ForPaw/controller/DTO/PostResponse.java
@@ -31,7 +31,7 @@ public class PostResponse {
                          String title,
                          String content,
                          LocalDateTime date,
-                         Integer answerNum) {}
+                         Long answerNum) {}
 
     public record PostImageDTO(Long id, String imageURL) {}
